src/test/java/LeetCode/MinimumTimetoCollectAllApplesinaTree_1443/SolutionTest.java
package LeetCode.MinimumTimetoCollectAllApplesinaTree_1443;

import LeetCode.Timer;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.List;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

class SolutionTest {

    Solution solution = new Solution();

    private static Stream<Arguments> testArguments() {
        return Stream.of(
                Arguments.of(7, new int[][]{{0, 1}, {0, 2}, {1, 4}, {1, 5}, {2, 3}, {2, 6}}, List.of(false, false, true, false, true, true, false), 8),
                Arguments.of(7, new int[][]{{0, 1}, {0, 2}, {1, 4}, {1, 5}, {2, 3}, {2, 6}}, List.of(false, false, true, false, false, true, false), 6),
                Arguments.of(7, new int[][]{{0, 1}, {0, 2}, {1, 4}, {1, 5}, {2, 3}, {2, 6}}, List.of(false, false, false, false, false, false, false), 0)
        );
    }

    @ParameterizedTest
    @MethodSource("testArguments")
    void test(int n, int[][] edges, List<Boolean> hasApple, int expected) {
        assertEquals(expected, solution.minTime(n, edges, hasApple));
    }
}